backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
anthropic==0.26.0
python-dotenv==1.0.1
httpx==0.27.0
pydantic==2.7.1
pytest==8.2.0
pytest-asyncio==0.23.7
